removed some redundancy wrt grouping notas in print view and approval view
